# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Suno AI API",
        "contact": {},
        "license": {},
        "version": "v1.0"
    },
    "paths": {
        "/api/v1/suno/createTask": {
            "post": {
                "description": "提交suno生成任务，返回音乐id等信息。",
                "produces": [
                    "application/json"
                ],
                "summary": "提交音乐生成任务",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "CreateGenerateSongTaskRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CreateGenerateSongTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/suno/getStatus": {
            "get": {
                "description": "通过音乐id,查询音乐生成状态",
                "produces": [
                    "application/json"
                ],
                "summary": "查询音乐生成状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "音乐id",
                        "name": "songId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.CreateGenerateSongTaskRequest": {
            "type": "object",
            "required": [
                "prompt",
                "tags",
                "title"
            ],
            "properties": {
                "callBack": {
                    "description": "回调地址",
                    "type": "string"
                },
                "prompt": {
                    "description": "歌词",
                    "type": "string"
                },
                "tags": {
                    "description": "风格、元素",
                    "type": "string"
                },
                "title": {
                    "description": "歌名",
                    "type": "string"
                }
            }
        },
        "controller.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}